chore: fix failing tests (#472)
