+++ src/domains/transaction/components/VoteDetail/VoteDetail.test.tsx
@@ -71,7 +71,7 @@ describe("VoteDetail", () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE),
+			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE_DELEGATE),
 		);
 
 		await expect(screen.findByText("Vote")).resolves.toBeVisible();
@@ -99,7 +99,7 @@ describe("VoteDetail", () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE),
+			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE_DELEGATE),
 		);
 
 		await expect(screen.findByText("Unvote")).resolves.toBeVisible();
@@ -127,7 +127,7 @@ describe("VoteDetail", () => {
 		);
 
 		await waitFor(() =>
-			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE),
+			expect(screen.getByTestId("Modal__inner")).toHaveTextContent(translations.MODAL_VOTE_DETAIL.TITLE_DELEGATE),
 		);
 
 		await expect(screen.findByText("Vote")).resolves.toBeVisible();
