+++ src/domains/transaction/cucumber/vote-transaction-multisig.ts
@@ -22,7 +22,7 @@ cucumber(
 		...preSteps,
 		"When she attempts to vote for a delegate with multisig wallet": async (t: TestController) => {
 			await t.click(Selector('[data-testid="AddressRow__select-2"]').withText(translations.COMMON.VOTE));
-			await t.expect(Selector("h2").withText(translations.VOTE.DELEGATE_TABLE.TITLE).exists).ok();
+			await t.expect(Selector("h2").withText(translations.VOTE.DELEGATE_TABLE.TITLE_DELEGATE).exists).ok();
 			await t.click(Selector('[data-testid="DelegateRow__toggle-0"]').withText(translations.COMMON.SELECT));
 			await t.expect(Selector("[data-testid=DelegateTable__footer]").exists).ok();
 			await t.click(
