+++ src/router/router.routes.test.ts
@@ -2,6 +2,6 @@ import { routes } from "./router.routes";
 
 describe("Router routes", () => {
 	it("should have routes", () => {
-		expect(routes).toHaveLength(32);
+		expect(routes).toHaveLength(33);
 	});
 });
