+++ src/tests/fixtures/coins/mainsail/devnet/blockchain.json
@@ -1,0 +1,9 @@
+{
+	"data": {
+		"block": {
+			"height": 564619,
+			"id": "70ba261e44084fae2d169785f672a31635d04ecdd19378fb4631aa3030f69035"
+		},
+		"supply": "12597803999999969"
+	}
+}
