+++ src/tests/fixtures/coins/mainsail/devnet/node-fees.json
@@ -1,0 +1,55 @@
+{
+	"data": {
+		"1": {
+			"transfer": {
+				"avg": "8133803",
+				"max": "0",
+				"min": "0",
+				"sum": "10000000"
+			},
+			"validatorRegistration": {
+				"avg": "977011494",
+				"max": "0",
+				"min": "0",
+				"sum": "2500000000"
+			},
+			"vote": {
+				"avg": "53508772",
+				"max": "0",
+				"min": "0",
+				"sum": "100000000"
+			},
+			"multiSignature": {
+				"avg": "500000000",
+				"max": "500000000",
+				"min": "500000000",
+				"sum": "500000000"
+			},
+			"multiPayment": {
+				"avg": "10000000",
+				"max": "10000000",
+				"min": "10000000",
+				"sum": "10000000"
+			},
+			"validatorResignation": {
+				"avg": "2500000000",
+				"max": "2500000000",
+				"min": "2500000000",
+				"sum": "2500000000"
+			},
+			"usernameRegistration": {
+				"avg": "1480769231",
+				"max": "0",
+				"min": "0",
+				"sum": "2500000000"
+			},
+			"usernameResignation": {
+				"avg": "2500000000",
+				"max": "2500000000",
+				"min": "2500000000",
+				"sum": "2500000000"
+			}
+		}
+	},
+	"meta": {}
+}
