+++ src/tests/fixtures/coins/mainsail/devnet/syncing.json
@@ -1,0 +1,8 @@
+{
+	"data": {
+		"blocks": -8498,
+		"height": 564585,
+		"id": 1,
+		"syncing": false
+	}
+}
